% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate}
\alias{evaluate}
\title{Evaluation of the classification of the probabilistic neural network.}
\usage{
evaluate(pnn)
}
\arguments{
\item{pnn}{Trained probabilistic neural network.(Required)}
}
\value{
pnn \code{list} with evaluated input and analysis charts
}
\description{
Evaluation of the classification of the probabilistic neural network.
}
\examples{
library(apnnClassifier)
data(trainData, testData)
pnn <- trainNeuralNet(train_set = trainData, test_set = testData, sigma = 0.5)
pnn <- evaluate(pnn)
View(pnn)
}
