% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stardardize.R
\name{standardize}
\alias{standardize}
\title{Data standardization}
\usage{
standardize(set, type = "punctual")
}
\arguments{
\item{set}{Data set. (Required)}

\item{type}{standardization type (Not required. Default value: "punctual". Valid values: "punctual" - "scale")}
}
\value{
standardized set.
}
\description{
Data standardization
}
\examples{
library(apnnClassifier)
data(testData)
Basic usage
testData <- standardize(testData)
For classification by minimum and maximum
testData <- standardize(testData, type = "scale")
View(testData)
}
