% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainNeuralNet.R
\name{trainNeuralNet}
\alias{trainNeuralNet}
\title{Neural network training of the pnn package for classification}
\usage{
trainNeuralNet(train_set, test_set, category_column = 1, sigma)
}
\arguments{
\item{train_set}{Training set. (Required)}

\item{test_set}{Testing set. (Required)}

\item{category_column}{Category column (Not required. Default value: 1)}

\item{sigma}{Optimum value for the activation function of the neural network. (Not required)}
}
\value{
pnn \code{list} trained with the classified testing set and network performance statistics.
}
\description{
Neural network training of the pnn package for classification
}
\examples{
library(apnnClassifier)
data(trainData, testData)
# Basic usage.
pnn <- trainNeuralNet(train_set = trainData, test_set = testData)
# If you know the approximate optimal value and the sorter column is not in the first position of the set.
pnn <- trainNeuralNet(train_set = trainData, test_set = testData, category_column =  *sorter column*, sigma = *sigma value*)
View(pnn)
}
